codeReview: 각 페이지의 데이터를 동기화
